requirements.txt
PyQt5
python-docx
pywin32
requests
